Move user-facing data sources to API package
Also: Add more default config parameters to data sources
