+++ spark-cypher-examples/src/main/scala/org/opencypher/spark/examples/MultipleGraphExample.scala
@@ -17,7 +17,7 @@ package org.opencypher.spark.examples
 
 import org.opencypher.okapi.api.graph.{GraphName, Namespace, QualifiedGraphName}
 import org.opencypher.spark.api.CAPSSession
-import org.opencypher.spark.impl.io.file.FileCsvPropertyGraphDataSource
+import org.opencypher.spark.api.io.file.FileCsvPropertyGraphDataSource
 
 /**
   * Demonstrates multiple graph capabilities by loading a social network from case class objects and a purchase network
