+++ spark-cypher/src/test/scala/org/opencypher/spark/impl/io/hdfs/HdfsCsvPropertyGraphDataSourceTest.scala
@@ -16,6 +16,7 @@
 package org.opencypher.spark.impl.io.hdfs
 
 import org.opencypher.okapi.api.graph.{GraphName, Namespace}
+import org.opencypher.spark.api.io.hdfs.HdfsCsvPropertyGraphDataSource
 import org.opencypher.spark.impl.CAPSConverters._
 import org.opencypher.spark.test.CAPSTestSuite
 import org.opencypher.spark.test.fixture.{MiniDFSClusterFixture, TeamDataFixture}
