+++ spark-cypher/src/test/scala/org/opencypher/spark/test/fixture/Neo4jServerFixture.scala
@@ -18,7 +18,7 @@ package org.opencypher.spark.test.fixture
 import org.neo4j.harness.{ServerControls, TestServerBuilders}
 import org.opencypher.okapi.test.BaseTestSuite
 import org.opencypher.okapi.test.fixture.BaseTestFixture
-import org.opencypher.spark.impl.io.neo4j.external.Neo4jConfig
+import org.opencypher.spark.api.io.neo4j.Neo4jConfig
 
 trait Neo4jServerFixture extends BaseTestFixture {
   self: SparkSessionFixture with BaseTestSuite =>
